// apps/web/src/contracts/generated/CoordWriteInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CoordColumn } from "./CoordColumn";
import type { CoordDelimiter } from "./CoordDelimiter";
import type { CoordPoint } from "./CoordPoint";
import type { TextEncoding } from "./TextEncoding";

export type CoordWriteInput = { points: Array<CoordPoint>, 
/**
 * Not `auto`.
 */
delimiter: CoordDelimiter, 
/**
 * Column order (`skip` is not written).
 */
columns: Array<CoordColumn>, 
/**
 * A first line with the column names.
 */
header: boolean, encoding: TextEncoding, };